diffusion_model/requirements.txt
torch 
torchvision
torchaudio
pandas
numpy
scikit-learn
biopython

